pyproject.toml
[project]
name = "timebox"
version = "0.6.0"
description = "Command line timeboxing utility for Things 3"
authors = [
    {name = "Walter Scott", email = "me@walteh.com"}
]
dependencies = [
    "rumps==0.4.0",
    "things.py==0.0.15",
    "watchdog>=3.0.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
